Make schema available for TypeScript items when using TypeScript NuGet
The schema for TypeScript items allows the IDE to display TypeScript items (.ts files) correctly in the hierarchy. The Web.ProjectSystem SDK targets imports the schema for TypeScript items as long as the TypeScript targets being referenced are not from the Microsoft.TypeScript.MSBuild NuGet package (that is, if TypeScript comes from the TS SDK instead). However, the TypeScript NuGet package doesn't actually contain the schema itself. So when a user installs the NuGet package in their project, the .ts files will disappear, This PR makes the reference to the schema unconditional.